+++ src/js/fetch.js
@@ -1,40 +1,14 @@
 import { createMarkUpCards } from './markup';
 import { createMarkUpIngridients } from './markupingridients';
-// export function getData(number) {
-//   const arrayOfPromise = [];
-//   for (let i = 1; i <= number; i++) {
-//     fetch('https://www.thecocktaildb.com/api/json/v1/1/random.php')
-//       .then(response => {
-//         if (!response.ok) {
-//           throw new Error(response.status);
-//         }
-//         console.log('response -  ', response);
-//         return response.json();
-//       })
-//       .then(data => {
-//         console.log('dtd -  ', data.drinks);
-//         console.log(createMarkUpCards(data.drinks));
-//       })
-//       .catch(error => {
-//         console.log('error   --->   ', error);
-//       });
-//   }
-// }
+import FetchService from './backend.js';
+
+const fetchService = new FetchService();
 
 //рандомное количество карточек
 export async function getCards(number) {
-  console.log('создаю карточки', number);
+  console.log('ЁЁЁЁЁ  создаю карточки', number);
   try {
-    const arrayPromise = [];
-    for (let i = 1; i <= number; i++) {
-      const response = await fetch(
-        'https://www.thecocktaildb.com/api/json/v1/1/random.php'
-      );
-      const data = await response.json();
-      arrayPromise.push(data);
-    }
-    const data = await Promise.all(arrayPromise);
-    const coctails = data.map(coctail => coctail.drinks[0]);
+    const coctails = await fetchService.randomCoctailsOnStart(number);
     createMarkUpCards(coctails, {
       add: true,
       h1Change: 'Cocktails',
@@ -51,14 +25,10 @@ export async function getCards(number) {
 export async function getCardsFirstLetter(key) {
   console.log('создаю карточки = ', key);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/search.php?f=${key}`
-    );
-    const drink = await response.json();
-
-    createMarkUpCards(drink.drinks, {
+    const response = await fetchService.byLetterCoctail(key);
+    createMarkUpCards(response, {
       add: false,
-      h1Change: `Searching results by first symbol ${key}`,
+      h1Change: 'Searching results',
     });
   } catch (error) {
     console.log(error.message);
@@ -72,13 +42,10 @@ export async function getCardsFirstLetter(key) {
 export async function getCardsByName(key) {
   console.log('создаю карточки = ', key);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/search.php?s=${key}`
-    );
-    const drink = await response.json();
-    createMarkUpCards(drink.drinks, {
+    const response = await fetchService.byNameCoctail(key);
+    createMarkUpCards(response, {
       add: false,
-      h1Change: `Searching results by string ${key}`,
+      h1Change: 'Searching results',
     });
   } catch (error) {
     console.log(error.message);
@@ -86,22 +53,13 @@ export async function getCardsByName(key) {
   }
 }
 // массив карточек: любиміе коктейли пользователя
-export async function getCardsByFavoritDrinks(arrayDrinks, h1 = 'Favorite cocktails') {
+export async function getCardsByFavoritDrinks(arrayDrinks) {
   console.log('arrayDrinks drinks = ', arrayDrinks);
   try {
-    const arrayPromise = [];
-    for (let i of arrayDrinks) {
-      const response = await fetch(
-        `https://www.thecocktaildb.com/api/json/v1/1/lookup.php?i=${i}`
-      );
-      const data = await response.json();
-      arrayPromise.push(data);
-    }
-    const data = await Promise.all(arrayPromise);
-    const coctails = data.map(coctail => coctail.drinks[0]);
+    const coctails = await fetchService.favoriteCoctailsById(arrayDrinks);
     createMarkUpCards(coctails, {
       add: false,
-      h1Change: h1,
+      h1Change: 'Favorite cocktails',
     });
   } catch (error) {
     console.log(error.message);
@@ -112,12 +70,8 @@ export async function getCardsByFavoritDrinks(arrayDrinks, h1 = 'Favorite cockta
 export async function getCardsByIngridient(key) {
   console.log('создаю карточки = ', key);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/search.php?i=${key}`
-    );
-    const drink = await response.json();
-    console.log(drink.ingredients);
-    createMarkUpIngridients(drink.ingredients, {
+    const response = fetchService.byIdCoctail(key);
+    createMarkUpIngridients(response, {
      display: 'modal',
     h1Change: '',
     });
@@ -130,19 +84,7 @@ export async function getCardsByIngridient(key) {
 export async function getCardsByFavoritIngridient(arrayIngridients) {
   console.log('arrayIngridient = ', arrayIngridients);
   try {
-    const arrayPromise = [];
-    for (let i of arrayIngridients) {
-      
-      const response = await fetch(
-        `https://www.thecocktaildb.com/api/json/v1/1/lookup.php?iid=${i}`
-      );
-      const data = await response.json();
-      arrayPromise.push(data);
-    }
-    const data = await Promise.all(arrayPromise);
-console.log(` await Promise.all(arrayPromise) `, data);
-    const ingridients = data.map(i => i.ingredients[0]);
-     console.log('ingridients = ', ingridients);
+    const ingridients = await fetchService.favoriteIngrById(arrayIngridients);
     createMarkUpIngridients(ingridients, {
       display: 'list',
       h1Change: 'Favorite ingredients',
@@ -153,19 +95,16 @@ console.log(` await Promise.all(arrayPromise) `, data);
   }
 }
 
+// -------------------------------------------------------------------------------------------------------------
+
 export async function getList(urlEnd, htmlElements) {
   console.log('создаю спиок по  = ', urlEnd);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/list.php?${urlEnd}`
-    );
-    const list = await response.json();
-    console.log(list.drinks);
-    htmlElements.innerHTML = list.drinks.map(
+    const list = await fetchService.fetchList(urlEnd);
+    console.log(list);
+    htmlElements.innerHTML = list.map(
       item =>
-        `<option value="${Object.values(item)[0]}">${
-          Object.values(item)[0]
-        }</option>`
+        `<option value="${Object.values(item)[0]}">${Object.values(item)[0]}</option>`
     ).join('');
   } catch (error) {
     console.log(error.message);
@@ -176,15 +115,10 @@ export async function getList(urlEnd, htmlElements) {
 export async function getCardsByСategory(category) {
   console.log('создаю карточки = ', category);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/filter.php?c=${category}`
-    );
-    const drink = await response.json();
-    console.log(drink.drinks);
-    const arrayID = drink.drinks.map(item => item.idDrink);
-     console.log(arrayID);
+    const response = await fetchService.fetchByCategory(category);
+    console.log('response', response);
      getCardsByFavoritDrinks(
-       arrayID,
+      response,
        `Searching results by category: ${category}`
      );
   } catch (error) {
@@ -196,14 +130,8 @@ export async function getCardsByСategory(category) {
 export async function getCardsByGlass(glass) {
   console.log('создаю карточки = ', glass);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/filter.php?g=${glass}`
-    );
-    const drink = await response.json();
-    console.log(drink.drinks);
-    const arrayID = drink.drinks.map(item => item.idDrink);
-    console.log(arrayID);
-    getCardsByFavoritDrinks(arrayID, `Searching results by glass: ${glass}`);
+    const response = await fetchService.fetchByGlass(glass);
+    getCardsByFavoritDrinks(response, `Searching results by glass: ${glass}`);
   } catch (error) {
     console.log(error.message);
     createMarkUpCards([], { add: false, h1Change: '' });
@@ -213,15 +141,9 @@ export async function getCardsByGlass(glass) {
 export async function getCardsByIngredient(Ingredient) {
   console.log('создаю карточки = ', Ingredient);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/filter.php?i=${Ingredient}`
-    );
-    const drink = await response.json();
-    console.log(drink.drinks);
-    const arrayID = drink.drinks.map(item => item.idDrink);
-    console.log(arrayID);
+    const response = await fetchService.fetchByIngr(Ingredient);
     getCardsByFavoritDrinks(
-      arrayID,
+      response,
       `Searching results by ingredient: ${Ingredient}`
     );
   } catch (error) {
@@ -233,15 +155,9 @@ export async function getCardsByIngredient(Ingredient) {
 export async function getCardsByselectByAlcoholic(Alcoholic) {
   console.log('создаю карточки = ', Alcoholic);
   try {
-    const response = await fetch(
-      `https://www.thecocktaildb.com/api/json/v1/1/filter.php?a=${Alcoholic}`
-    );
-    const drink = await response.json();
-    console.log(drink.drinks);
-    const arrayID = drink.drinks.map(item => item.idDrink);
-    console.log(arrayID);
+    const response = await fetchService.fetchByAlcohol(Alcoholic);
     getCardsByFavoritDrinks(
-      arrayID,
+      response,
       `Searching results by alcoholic: ${Alcoholic}`
     );
   } catch (error) {
