+++ src/js/fetch_old.js
@@ -1,14 +1,40 @@
 import { createMarkUpCards } from './markup';
 import { createMarkUpIngridients } from './markupingridients';
-import FetchService from './backend.js';
-
-const fetchService = new FetchService();
+// export function getData(number) {
+//   const arrayOfPromise = [];
+//   for (let i = 1; i <= number; i++) {
+//     fetch('https://www.thecocktaildb.com/api/json/v1/1/random.php')
+//       .then(response => {
+//         if (!response.ok) {
+//           throw new Error(response.status);
+//         }
+//         console.log('response -  ', response);
+//         return response.json();
+//       })
+//       .then(data => {
+//         console.log('dtd -  ', data.drinks);
+//         console.log(createMarkUpCards(data.drinks));
+//       })
+//       .catch(error => {
+//         console.log('error   --->   ', error);
+//       });
+//   }
+// }
 
 //рандомное количество карточек
 export async function getCards(number) {
-  console.log('ЁЁЁЁЁ  создаю карточки', number);
+  console.log('создаю карточки', number);
   try {
-    const coctails = await fetchService.randomCoctailsOnStart(number);
+    const arrayPromise = [];
+    for (let i = 1; i <= number; i++) {
+      const response = await fetch(
+        'https://www.thecocktaildb.com/api/json/v1/1/random.php'
+      );
+      const data = await response.json();
+      arrayPromise.push(data);
+    }
+    const data = await Promise.all(arrayPromise);
+    const coctails = data.map(coctail => coctail.drinks[0]);
     createMarkUpCards(coctails, {
       add: true,
       h1Change: 'Cocktails',
@@ -25,10 +51,14 @@ export async function getCards(number) {
 export async function getCardsFirstLetter(key) {
   console.log('создаю карточки = ', key);
   try {
-    const response = await fetchService.byLetterCoctail(key);
-    createMarkUpCards(response, {
+    const response = await fetch(
+      `https://www.thecocktaildb.com/api/json/v1/1/search.php?f=${key}`
+    );
+    const drink = await response.json();
+
+    createMarkUpCards(drink.drinks, {
       add: false,
-      h1Change: 'Searching results',
+      h1Change: `Searching results by first symbol ${key}`,
     });
   } catch (error) {
     console.log(error.message);
@@ -42,10 +72,13 @@ export async function getCardsFirstLetter(key) {
 export async function getCardsByName(key) {
   console.log('создаю карточки = ', key);
   try {
-    const response = await fetchService.byNameCoctail(key);
-    createMarkUpCards(response, {
+    const response = await fetch(
+      `https://www.thecocktaildb.com/api/json/v1/1/search.php?s=${key}`
+    );
+    const drink = await response.json();
+    createMarkUpCards(drink.drinks, {
       add: false,
-      h1Change: 'Searching results',
+      h1Change: `Searching results by string ${key}`,
     });
   } catch (error) {
     console.log(error.message);
@@ -53,13 +86,22 @@ export async function getCardsByName(key) {
   }
 }
 // массив карточек: любиміе коктейли пользователя
-export async function getCardsByFavoritDrinks(arrayDrinks) {
+export async function getCardsByFavoritDrinks(arrayDrinks, h1 = 'Favorite cocktails') {
   console.log('arrayDrinks drinks = ', arrayDrinks);
   try {
-    const coctails = await fetchService.favoriteCoctailsById(arrayDrinks);
+    const arrayPromise = [];
+    for (let i of arrayDrinks) {
+      const response = await fetch(
+        `https://www.thecocktaildb.com/api/json/v1/1/lookup.php?i=${i}`
+      );
+      const data = await response.json();
+      arrayPromise.push(data);
+    }
+    const data = await Promise.all(arrayPromise);
+    const coctails = data.map(coctail => coctail.drinks[0]);
     createMarkUpCards(coctails, {
       add: false,
-      h1Change: 'Favorite cocktails',
+      h1Change: h1,
     });
   } catch (error) {
     console.log(error.message);
@@ -70,8 +112,12 @@ export async function getCardsByFavoritDrinks(arrayDrinks) {
 export async function getCardsByIngridient(key) {
   console.log('создаю карточки = ', key);
   try {
-    const response = fetchService.byIdCoctail(key);
-    createMarkUpIngridients(response, {
+    const response = await fetch(
+      `https://www.thecocktaildb.com/api/json/v1/1/search.php?i=${key}`
+    );
+    const drink = await response.json();
+    console.log(drink.ingredients);
+    createMarkUpIngridients(drink.ingredients, {
      display: 'modal',
     h1Change: '',
     });
@@ -84,7 +130,19 @@ export async function getCardsByIngridient(key) {
 export async function getCardsByFavoritIngridient(arrayIngridients) {
   console.log('arrayIngridient = ', arrayIngridients);
   try {
-    const ingridients = await fetchService.favoriteIngrById(arrayIngridients);
+    const arrayPromise = [];
+    for (let i of arrayIngridients) {
+      
+      const response = await fetch(
+        `https://www.thecocktaildb.com/api/json/v1/1/lookup.php?iid=${i}`
+      );
+      const data = await response.json();
+      arrayPromise.push(data);
+    }
+    const data = await Promise.all(arrayPromise);
+console.log(` await Promise.all(arrayPromise) `, data);
+    const ingridients = data.map(i => i.ingredients[0]);
+     console.log('ingridients = ', ingridients);
     createMarkUpIngridients(ingridients, {
       display: 'list',
       h1Change: 'Favorite ingredients',
@@ -94,8 +152,6 @@ export async function getCardsByFavoritIngridient(arrayIngridients) {
     createMarkUpIngridients([], { add: false, h1Change: '' });
   }
 }
-
-// -------------------------------------------------------------------------------------------------------------
 
 export async function getList(urlEnd, htmlElements) {
   console.log('создаю спиок по  = ', urlEnd);
